fix(reorderSiteAwards): shrink rows and add autoscroll
